requirements.txt
pyaudio
openai
spotipy
anthropic
pygame
pytest
websocket-client
elevenlabs